refactoring of virtual memory module
