+++ kernel/src/library/memory/physical.rs
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright 2022 The unCORE Kernel Organization
 
-use crate::library::architectures::memory::physical as architecture_physical_memory;
+use crate::prelude::*;
+use crate::library::architectures::memory as architecture_memory;
 
 /// ### Global Kernel Frame Allocator
 ///
 /// Structure containing the kernel's global frame allocator.
-pub(super) static mut KERNEL_FRAME_ALLOCATOR: spin::once::Once<architecture_physical_memory::FrameAllocator> =
+pub(super) static mut KERNEL_FRAME_ALLOCATOR: spin::once::Once<architecture_memory::FrameAllocator> =
 	spin::once::Once::new();
 
 /// ### Acquire A Mutable Frame Allocator Reference
@@ -18,86 +19,11 @@ pub(super) static mut KERNEL_FRAME_ALLOCATOR: spin::once::Once<architecture_phys
 ///
 /// This function is `unsafe` as it returns a mutable reference from a `static mut`
 /// variable. This function will panic if the allocator is not initialized.
-pub(crate) unsafe fn get_frame_allocator() -> &'static mut architecture_physical_memory::FrameAllocator
+pub(crate) unsafe fn get_frame_allocator() -> &'static mut architecture_memory::FrameAllocator
 {
 	KERNEL_FRAME_ALLOCATOR
 		.get_mut()
 		.expect("Could not acquire frame allocator (is it initialized?)")
-}
-
-/// ### A Physical Address Abstraction
-///
-/// This is an opaque wrapper type that contains the address as its first type.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct PhysicalAddress(usize);
-
-impl PhysicalAddress
-{
-	/// ### Create a New Physical Address
-	///
-	/// Constructs a new physical address.
-	pub fn new(address: impl Into<usize>) -> Self { Self(address.into()) }
-
-	/// ### Get the Inner Value
-	///
-	/// Returns the inner value, i.e. content that is wrapped by this type.
-	pub fn inner(&self) -> usize { self.0 }
-}
-
-impl ::core::ops::Add for PhysicalAddress
-{
-	type Output = Self;
-
-	fn add(self, rhs: Self) -> Self::Output { Self::new(self.inner() + rhs.inner()) }
-}
-
-impl ::core::ops::Add<usize> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn add(self, rhs: usize) -> Self::Output { Self::new(self.inner() + rhs) }
-}
-
-impl ::core::ops::Add<u64> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn add(self, rhs: u64) -> Self::Output { Self::new(self.inner() + rhs as usize) }
-}
-
-impl ::core::ops::Add<i64> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn add(self, rhs: i64) -> Self::Output { Self::new(self.inner() + rhs as usize) }
-}
-
-impl ::core::ops::Sub for PhysicalAddress
-{
-	type Output = Self;
-
-	fn sub(self, rhs: Self) -> Self::Output { Self::new(self.inner() - rhs.inner()) }
-}
-
-impl ::core::ops::Sub<usize> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn sub(self, rhs: usize) -> Self::Output { Self::new(self.inner() - rhs) }
-}
-
-impl ::core::ops::Sub<u64> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn sub(self, rhs: u64) -> Self::Output { Self::new(self.inner() - rhs as usize) }
-}
-
-impl ::core::ops::Sub<i64> for PhysicalAddress
-{
-	type Output = Self;
-
-	fn sub(self, rhs: i64) -> Self::Output { Self::new(self.inner() - rhs as usize) }
 }
 
 /// ### Representation of a Page
@@ -107,7 +33,7 @@ impl ::core::ops::Sub<i64> for PhysicalAddress
 pub struct Frame<S: super::ChunkSize = super::ChunkSizeDefault>
 {
 	/// Where the frame starts in physical memory.
-	start_address: PhysicalAddress,
+	start_address: memory::PhysicalAddress,
 	/// How big the physical frame is.
 	size:          ::core::marker::PhantomData<S>,
 }
@@ -117,7 +43,7 @@ impl<S: super::ChunkSize> Frame<S>
 	/// ### Create a New Frame
 	///
 	/// Creates a new physical frame instance.
-	pub fn new(start_address: PhysicalAddress) -> Self
+	pub fn new(start_address: memory::PhysicalAddress) -> Self
 	{
 		Self {
 			start_address,
@@ -128,7 +54,7 @@ impl<S: super::ChunkSize> Frame<S>
 	/// ### Start Address of a Frame
 	///
 	/// Returns the starts address of the given frame.
-	pub fn start(&self) -> PhysicalAddress { self.start_address }
+	pub fn start(&self) -> memory::PhysicalAddress { self.start_address }
 }
 
 /// ### Capability of Allocating Frames
